build(analyzers): apply the canonical analyzer baseline
The repository built with the analyzers muted: Directory.Build.props set no
TreatWarningsAsErrors, EnforceCodeStyleInBuild or AnalysisMode, so the style
rules .editorconfig declared reached the IDE and never the build, and the
"0 warnings" gate measured compiler warnings only. With the gate on and no
repository nuget.config, restore failed first on NU1507 from the machine's
package sources.

The tree now carries the canonical .editorconfig and Directory.Build.props
from the project-layout skill, with the project-specific groups kept and two
rules muted with a reason: IDE0045 and IDE0046 rewrite a guard clause into
nested conditionals, which is not the house shape, and CA1001 is off for
test doubles whose lifetime ends with the test. A root nuget.config clears
the inherited sources and names nuget.org alone.

The 172 warning sites the baseline exposed are cleared: collection
expressions, PascalCase constants, culture-explicit formatting of UI text
with CurrentCulture, auto-properties, static members, primary constructors,
and a PostgresDataProvider that owns its fault subject and completes it on
Dispose. The 26 files under SemiPlot.Tests.Integration/Integration/ move to
the project root, where their namespace already places them.

CI enforces both halves: the Release build with TreatWarningsAsErrors and,
on the Linux job, dotnet format --verify-no-changes for the format-only
rules no hook runs there.

+++ SemiPlot/SemiPlot.Tests.Unit/Postgres/PenLineStyleTests.cs
@@ -16,7 +16,7 @@ public sealed class PenLineStyleTests
 	[Fact]
 	public void TheOrdinalsAreTheValuesTheCatalogueStores()
 	{
-		((short)PenLineStyle.Interpolated).Should().Be((short)0);
-		((short)PenLineStyle.Stepped).Should().Be((short)1);
+		((short)PenLineStyle.Interpolated).Should().Be(0);
+		((short)PenLineStyle.Stepped).Should().Be(1);
 	}
 }
